refactor pending rides
